+++ lib/eex/test/test_helper.exs
@@ -1,1 +1,1 @@
-ExUnit.start [debug: "--debug" in System.argv]
+ExUnit.start [trace: "--trace" in System.argv]