+++ lib/mix/test/test_helper.exs
@@ -1,7 +1,7 @@
 Mix.start()
 Mix.shell(Mix.Shell.Process)
 
-ExUnit.start [debug: "--debug" in System.argv]
+ExUnit.start [trace: "--trace" in System.argv]
 System.put_env("EXUNIT_CONFIG", "none")
 
 defmodule MixTest.Case do